mapeo de base de datos bidireccional

+++ src/main/java/com/zabdev/biblioteca/repositorios/BibliotecaRepository.java
@@ -3,6 +3,6 @@ package com.zabdev.biblioteca.repositorios;
 import com.zabdev.biblioteca.entidades.Biblioteca;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BibliotecaRepository extends JpaRepository<Biblioteca, Long> {
+public interface BibliotecaRepository extends JpaRepository<Biblioteca, Integer> {
 
 }
